Update bm.c
this will add a Gamma frequency to the Brain Machine.  Mitch Altman was kind enough to send it to me.  When mice were exposed to light flashed at 40Hz, within an otherwise dark room, they had half the brain destroying amyloid plaque of control mice. Check out Bringing Gamma Back on https://www.wnycstudios.org/story/bringing-gamma-back.  When light and sound are both pulsed at  40 hz, plaque is removed not only from the visual and auditory cortex, but from the hippocampus itself.  And, apparently it works on humans too, Nothing has ever reversed Alzheimer's damage before.  Vielight uses NearInfraRed light on the head and clipped to the inside of the nose. Amazing results after 6 weeks in the viability study, but their intranasal is, for reasons I cannot understand, at 10Hz, and they are counting on NearInfraRed to penetrate all the way down through the forebrain and the midbrain to get to the hippocampus. This should do what Vielight does and more. Next we just need a 40Hz lollipop, in nearinfrared, and paired with pulsed Ultrasound.  While rare, there is a genetic form of Alzheimer's that starts hitting at 40, at 50 your brain starts retiring, at 60 your in a home, at 65 your dead.  This will really help those people, a few of whom are my relatives.